Basic month view display (not yet tested)
Added calendar_month.html template and corresponding url in urls.py and
views in views.py. The calendar html is generated by the MonthView view
and Calendar class by overriding HTMLCalendar defaults. This version
does not yet have the ability to view previous and next months. This
version also has not yet been tested because of difficulties with
creating task objects in the admin panel for testing.