Update templates, views, and tests for month view
Updated calendar_month.html and base_template.html to include blocks for
custom CSS styling. Updated Calendar class in views to have buttons
(which are used to display tasks in the calendar) span the width of the
day cells; the custom CSS in calendar_month.html prevents the overflow
text from escaping the cells.

I modified the Task model to use a CustomUser instead of a User in order
to be able to create Tasks, which requires a valid (Custom)User to be set;
I could not figure out how to make my integration tests work without
using a CustomUser.

The tests I added for the month view are one to check if the URL for the
month view properly brings up the calendar_month template via MonthView
and an integration test to check for if created tasks are displayed within
the generated calendar month.

I also made small adjustments to spacing between classes and whatnot as
well as added comments for readability.